requirements.txt
psycopg2
requests
simplejson
aiven.client
configparser
pyvis
networkx
dash
plotly